+++ CHANGELOG.md
@@ -2,10 +2,14 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+### 1.2.0 (2020-05-17)
+
+changed the code in bulkRequest to use `endpointName` feature of `@kintone/rest-api-client@1.2.0`
+
 ### 1.1.2 (2020-02-28)
 
 fix: app in guest space
 
 ### 1.1.1 (2020-02-28)
 
-use `@kintone/rest-api-client`
+changed the whole code to use `@kintone/rest-api-client`